# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Motionclash
  version: 1.0.0
servers:
- url: https://localhost:7000
paths:
...
  /api/score/{id}:
    delete:
      operationId: deleteScoreById
      parameters:
      - description: "ID of the Score to retrieve, update or delete."
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Score deleted
        "404":
          description: Score not found
      summary: Delete a Score by ID.
      tags:
      - Score
      x-accepts: application/json
      x-tags:
      - tag: Score
    get:
      operationId: getScoreById
      parameters:
      - description: "ID of the Score to retrieve, update or delete."
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Score'
          description: Score object
        "404":
          description: Score not found
      summary: Get a Score by ID.
      tags:
      - Score
      x-accepts: application/json
      x-tags:
      - tag: Score
    put:
      operationId: upsertScoreById
      parameters:
      - description: "ID of the Score to retrieve, update or delete."
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Score'
        description: Score object to upsert.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Score'
          description: Score object
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Score'
          description: Score created
        "404":
          description: Score not found
      summary: Upsert a Score by ID.
      tags:
      - Score
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Score
  /api/score:
    get:
      operationId: getAllScores
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Score'
                type: array
          description: List of Score objects
      summary: Get all Scores.
      tags:
      - Score
      x-accepts: application/json
      x-tags:
      - tag: Score
    post:
      operationId: createScore
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Score'
        description: Score object to create.
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Score'
          description: Score created
      summary: Create a new Score.
      tags:
      - Score
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Score
  /api/score/race/{id}:
    get:
      operationId: getAllScoresOfRace
      parameters:
      - description: Id of the race to retrive scores for.
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Score'
                type: array
          description: List of Score objects in a race
      summary: Get all Scores of a race.
      tags:
      - Score
      x-accepts: application/json
      x-tags:
      - tag: Score
  /api/score/group/{id}:
    get:
      operationId: getScoreOfGroup
      parameters:
      - description: Id of the group to retrive score for.
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Score'
          description: List of Score objects
      summary: Get score of a group
      tags:
      - Score
      x-accepts: application/json
      x-tags:
      - tag: Score
    post:
      operationId: WaypointReached
      parameters:
      - description: Id of the group
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Waypoint'
        description: Waypoint object to complete.
        required: true
      responses:
        "200":
          description: Result of the operation
      summary: The given group reached a new waypoint
      tags:
      - Score
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Score
...
components:
  schemas:
...
    Score:
      example:
        raceId: 123e4567-e89b-12d3-a456-426614174000
        RoutePoints:
        - null
        - null
        groupId: 123e4567-e89b-12d3-a456-426614174000
        id: 123e4567-e89b-12d3-a456-426614174000
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        raceId:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        groupId:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        RoutePoints:
          items:
            $ref: '#/components/schemas/ActiveWaypoint'
          type: array
      type: object
...
    Waypoint:
      example:
        latitiude: 0.8008281904610115
        longitude: 6.027456183070403
      properties:
        latitiude:
          type: number
        longitude:
          type: number
      type: object
    ActiveWaypoint:
      allOf:
      - $ref: '#/components/schemas/Waypoint'
      - $ref: '#/components/schemas/ActiveWaypoint_allOf'
    ActiveWaypoint_allOf:
      properties:
        EstimatedTime:
          type: string
        ReachedTime:
          type: string
      type: object
      example: null